+++ src/main/java/com/prgrms/needit/domain/user/login/controller/UserController.java
@@ -1,11 +1,11 @@
-package com.prgrms.needit.domain.user.controller;
+package com.prgrms.needit.domain.user.login.controller;
 
 import com.prgrms.needit.common.response.ApiResponse;
-import com.prgrms.needit.domain.user.dto.LoginRequest;
-import com.prgrms.needit.domain.user.dto.TokenResponse;
-import com.prgrms.needit.domain.user.service.UserService;
+import com.prgrms.needit.domain.user.login.dto.LoginRequest;
+import com.prgrms.needit.domain.user.login.dto.TokenResponse;
+import com.prgrms.needit.domain.user.login.service.UserService;
+import javax.validation.Valid;
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -23,7 +23,7 @@ public class UserController {
 
 	@PostMapping("/login")
 	public ResponseEntity<ApiResponse<TokenResponse>> login(
-		@Validated @RequestBody LoginRequest login
+		@Valid @RequestBody LoginRequest login
 	) {
 		return ResponseEntity.ok(ApiResponse.of(userService.login(login)));
 	}
