+++ src/main/java/com/prgrms/needit/domain/user/login/service/UserService.java
@@ -1,4 +1,4 @@
-package com.prgrms.needit.domain.user.service;
+package com.prgrms.needit.domain.user.login.service;
 
 import com.prgrms.needit.common.config.jwt.JwtTokenProvider;
 import com.prgrms.needit.common.enums.UserType;
@@ -8,8 +8,8 @@ import com.prgrms.needit.domain.center.entity.Center;
 import com.prgrms.needit.domain.center.repository.CenterRepository;
 import com.prgrms.needit.domain.member.entity.Member;
 import com.prgrms.needit.domain.member.repository.MemberRepository;
-import com.prgrms.needit.domain.user.dto.LoginRequest;
-import com.prgrms.needit.domain.user.dto.TokenResponse;
+import com.prgrms.needit.domain.user.login.dto.LoginRequest;
+import com.prgrms.needit.domain.user.login.dto.TokenResponse;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
